pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "surreality-auth"
version = "1.0.0"
description = "Shared authentication middleware for Surreality services"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "SurrealityAI", email = "engineering@surreality.ai"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "fastapi>=0.68.0",
    "PyJWT>=2.0.0",
    "supabase>=1.0.0",
    "python-dotenv>=0.19.0",
]

[project.urls]
"Homepage" = "https://github.com/SurrealityAI/surreality-auth"
"Bug Reports" = "https://github.com/SurrealityAI/surreality-auth/issues"
"Source" = "https://github.com/SurrealityAI/surreality-auth"